+++ src/Tools/dotnet-counters/Program.cs
@@ -22,7 +22,7 @@ namespace Microsoft.Diagnostics.Tools.Counters
                 "Start monitoring a .NET application", 
                 new Option[] { ProcessIdOption(), RefreshIntervalOption() },
                 argument: CounterList(),
-                handler: CommandHandler.Create<CancellationToken, string, IConsole, int, float>(new CounterMonitor().Monitor));
+                handler: CommandHandler.Create<CancellationToken, string, IConsole, int, int>(new CounterMonitor().Monitor));
 
         private static Option ProcessIdOption() =>
             new Option(
@@ -34,7 +34,7 @@ namespace Microsoft.Diagnostics.Tools.Counters
             new Option(
                 new[] { "-r", "--interval" }, 
                 "The number of seconds to delay between updating the displayed counters.",
-                new Argument<float> { Name = "interval" });
+                new Argument<int> { Name = "interval" });
 
         private static Argument CounterList() =>
             new Argument<string> {
